Extract LogProxy concern from CommitsController
Lifts the upstream-log fetcher + HEAD probe + the LogNotFound /
LogTooLarge / LogFetchError error types out of CommitsController
and into a controllers/concerns/log_proxy.rb concern, along with
the byte cap and timeout constants. CommitsController#build_log
and #build_log_status become thin wrappers that just shape the
per-action error messages and cache key — same behavior as
before.

The motivation is that the test-on-commit page is about to grow
its own log proxy (for the three per-test files at
mesa-logs.flatironinstitute.org/<sha>/<computer>/<test>/{out,mk,err}.txt),
and re-implementing the byte-cap-streaming-fetch + HEAD-probe-
with-cached-result dance in two controllers would diverge over
time.
